I have color coded each player. I also made it so the sharks kill you now.
There are two tasks left that I'd like to do for tomorrow and something else for the victory screen
if possible.

I'd like to create a zooming camera such that the players are never off screen unless they're too far apart, and possibly
have a boundary at that location which causes you to bounce back. Having this boundary, I'd also like to expand the number
of possible fish and sharks in the game.

The second feature I want to put in is a fade between the title screen and the gameplay screen as there's a really cool way
we could do that.

These should both be doable later today when I wake up.

Let me know if you plan on working on one of these.