Fixed the intermittent dash bug. It seems to be caused by the check inside applyMovement where the players state is state.Dash. I removed uK's usage there and made the case equivalent to the second case and the bug is no longer visible.
Fixed the bug where sometimes the player would disappear from changing from big to medium or medium to small. It was caused by inconsistent Animation usage
in the Penguin class as well as a logic error in the animation's frame advancement logic.

I also readded calorie display on the top because when the screen is zoomed out it's pretty much impossible to tell how much health you have.